backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.9
pyjwt[crypto]==2.10.1
httpx==0.28.1
supabase==2.11.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
celery==5.4.0
redis==5.2.1
anthropic==0.42.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
